requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
insightface==0.7.3
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
onnxruntime==1.19.2
pydantic==2.9.2
